+++ CHANGELOG.md
@@ -1,0 +1,6 @@
+Version History
+====
+    * All Version bumps are required to update this file as well!!
+----
+
+* 0.5.0 - Auto-inject Client Helper into Views/Controllers. Pass through to Snippet tag entire set of HTML attributes. Implement new namespacing and support Bundler.require.
