+++ lib/cortex-snippets-client-ruby.rb
@@ -1,0 +1,2 @@
+require 'cortex/snippets/client'
+require 'cortex/snippets/client/version'
