--- lib/cortex/snippets-client.rb
@@ -1,4 +1,0 @@
-require 'cortex/snippets/version'
-require 'cortex/snippets/client'
-require 'cortex/snippets/view_helpers'
-require 'cortex/snippets/railtie' if defined?(Rails)
