+++ lib/cortex/snippets/client/version.rb
@@ -1,0 +1,7 @@
+module Cortex
+  module Snippets
+    module Client
+      VERSION = '0.5.0'
+    end
+  end
+end
