--- lib/cortex/snippets/version.rb
@@ -1,5 +1,0 @@
-module Cortex
-  module Snippets
-    VERSION = '0.4.4'
-  end
-end
