+++ spec/cortex/snippets_client_spec.rb
@@ -1,7 +1,7 @@
 require 'spec_helper'
 
-describe Cortex::Snippets do
+describe Cortex::Snippets::Client do
   it 'has a version number' do
-    expect(Cortex::Snippets::VERSION).not_to be nil
+    expect(Cortex::Snippets::Client::VERSION).not_to be nil
   end
 end
