+++ spec/spec_helper.rb
@@ -1,2 +1,2 @@
 $LOAD_PATH.unshift File.expand_path('../../lib', __FILE__)
-require 'cortex/snippets-client'
+require 'cortex-snippets-client-ruby'
